BigDecimal

+++ Station/src/main/java/com/fice/resource/StationResource.java
@@ -7,6 +7,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
+
 @RestController
 public class StationResource {
 
@@ -14,7 +16,7 @@ public class StationResource {
     StationService stationService;
 
     @GetMapping(path = "/stations/{sourceStation}/{destinationStation}", produces = MediaType.TEXT_PLAIN_VALUE)
-    public double checkRouteResource(@PathVariable("sourceStation") String sourceStation, @PathVariable("destinationStation") String destinationStation) {
+    public BigDecimal checkRouteResource(@PathVariable("sourceStation") String sourceStation, @PathVariable("destinationStation") String destinationStation) {
         return stationService.checkRoute(sourceStation,destinationStation);
     }
